Creating separate file for AppFile
