+++ client/tests/unit/AppContentTests.cpp
@@ -108,85 +108,6 @@ std::unique_ptr<AppContent> GetAppContent(const std::string& contentNameSpace,
 }
 } // namespace
 
-TEST("Testing AppFile::Make()")
-{
-    const std::string fileId{"myFileId"};
-    const std::string url{"myUrl"};
-    const uint64_t sizeInBytes{1234};
-    const std::unordered_map<HashType, std::string> hashes{{HashType::Sha1, "mySha1"}, {HashType::Sha256, "mySha256"}};
-    const std::vector<Architecture> architectures{Architecture::Amd64};
-    const std::vector<std::string> platformApplicabilityForPackage{"myPlatformApplicabilityForPackage"};
-    const std::string fileMoniker{"myFileMoniker"};
-
-    const std::unique_ptr<AppFile> file =
-        GetAppFile(fileId, url, sizeInBytes, hashes, architectures, platformApplicabilityForPackage, fileMoniker);
-
-    CHECK(fileId == file->GetFileId());
-    CHECK(url == file->GetUrl());
-    CHECK(sizeInBytes == file->GetSizeInBytes());
-    CHECK(hashes == file->GetHashes());
-
-    SECTION("Testing File equality operators")
-    {
-        SECTION("Equal")
-        {
-            auto CompareFileEqual = [&file](const std::unique_ptr<AppFile>& sameFile) {
-                REQUIRE((*file == *sameFile));
-                REQUIRE_FALSE((*file != *sameFile));
-            };
-
-            CompareFileEqual(GetAppFile(fileId,
-                                        url,
-                                        sizeInBytes,
-                                        hashes,
-                                        architectures,
-                                        platformApplicabilityForPackage,
-                                        fileMoniker));
-        }
-
-        SECTION("Not equal")
-        {
-            auto CompareFileNotEqual = [&file](const std::unique_ptr<AppFile>& otherFile) {
-                REQUIRE((*file != *otherFile));
-                REQUIRE_FALSE((*file == *otherFile));
-            };
-
-            CompareFileNotEqual(
-                GetAppFile("", url, sizeInBytes, hashes, architectures, platformApplicabilityForPackage, fileMoniker));
-            CompareFileNotEqual(GetAppFile(fileId,
-                                           "",
-                                           sizeInBytes,
-                                           hashes,
-                                           architectures,
-                                           platformApplicabilityForPackage,
-                                           fileMoniker));
-            CompareFileNotEqual(
-                GetAppFile(fileId, url, 0, hashes, architectures, platformApplicabilityForPackage, fileMoniker));
-            CompareFileNotEqual(
-                GetAppFile(fileId, url, sizeInBytes, {}, architectures, platformApplicabilityForPackage, fileMoniker));
-            CompareFileNotEqual(
-                GetAppFile(fileId, url, sizeInBytes, hashes, {}, platformApplicabilityForPackage, fileMoniker));
-            CompareFileNotEqual(GetAppFile(fileId, url, sizeInBytes, hashes, architectures, {}, fileMoniker));
-            CompareFileNotEqual(GetAppFile(fileId, url, sizeInBytes, hashes, {}, platformApplicabilityForPackage, {}));
-            CompareFileNotEqual(GetAppFile("", "", 0, {}, architectures, platformApplicabilityForPackage, fileMoniker));
-            CompareFileNotEqual(GetAppFile("MYFILEID",
-                                           url,
-                                           sizeInBytes,
-                                           hashes,
-                                           architectures,
-                                           platformApplicabilityForPackage,
-                                           fileMoniker));
-            CompareFileNotEqual(GetAppFile(fileId,
-                                           "MYURL",
-                                           sizeInBytes,
-                                           hashes,
-                                           architectures,
-                                           platformApplicabilityForPackage,
-                                           fileMoniker));
-        }
-    }
-}
-
 TEST("Testing AppContent::Make()")
 {
     const std::string contentNameSpace{"myNameSpace"};
